+++ tango/__init__.py
@@ -2,6 +2,20 @@
 A Python library for choreographing your machine learning research.
 """
 
+__all__ = [
+    "Format",
+    "DillFormat",
+    "DillFormatIterator",
+    "JsonFormat",
+    "JsonFormatIterator",
+    "Step",
+    "step_graph_from_params",
+    "tango_dry_run",
+    "StepCache",
+    "MemoryStepCache",
+    "DirectoryStepCache",
+]
+
 from .format import Format, DillFormat, DillFormatIterator, JsonFormat, JsonFormatIterator
 from .step import Step, step_graph_from_params, tango_dry_run
 from .step_cache import StepCache, MemoryStepCache, DirectoryStepCache