+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - `tango run` command will acquire a lock on the directory to avoid race conditions.
+- Integrations can now be installed with `pip install tango[INTEGRATION_NAME]`. For example,
+  `pip install tango[torch]`.
 
 ## [v0.0.3](https://github.com/allenai/tango/releases/tag/v0.0.3) - 2021-09-27